requirements.txt
Requests
python-dotenv
rich
gpt4all
openai
transformers
selenium
webdriver_manager